Module Api

module Dep : sig ... end

module Env : sig ... end

module Errors : sig ... end

Errors handling

module Files : sig ... end

module Meta : sig ... end

The dkmeta library is an extension of dedukti that allows the user to normalize terms according to meta rules. Meta rules are written in the dedukti syntax. dkmeta also offers a way to reify the syntax of dedukti in dedukti. This allows you to get around the limitations of the rewrite engine offers par Dedukti: you can rewrite products, static symbols... The overhead introduced by this reification is linear if you choose the LF encoding. However, you might create your own encoding for your own goals.

module Pp : sig ... end

module Processor : sig ... end

A process module processes entries. It is parameterized by an environment.