Module Api.Pp

val print_db_enabled : bool Stdlib.ref

This function is here for debug purposes. It prints terms with De Bruijn indices. The terms output this way cannot be parsed again.

val print_default_name : bool Stdlib.ref

Names for rules are optional. If this option is set to true, the printing functions will output default rule names given by Dedukti.

val print_module_name : bool Stdlib.ref

module type Sig = sig ... end

module type Printer = sig ... end

module Make (S : Sig) : Printer

module Default : Printer