Module type Processor.Interface

type 'a t

val handle_input : Parsers.Parser.input -> ?hook:hook -> 'a t -> 'a

handle_input input hook processor applies the processor processor on the input. hook.hook_before is executed once before the processor and hook.hook_after is executed once after the processor. By default (without hooks), if an exception exn has been raised while processing the data it is raised at top-level.

val handle_files : string list -> ?hook:hook -> 'a t -> 'a

handle_files files hook processor apply a processor on each file of files. hook is used once by file. The result is the one given once each file has been processed.

val fold_files : string list -> ?hook:hook -> f:('a -> 'b -> 'b) -> default:'b -> 'a t -> 'b