Module type Processor.S

The actual type of the processor as a module

type t

result type of the processor

val handle_entry : Env.t -> Parsers.Entry.entry -> unit

handle_entry env entry processed the entry entry in the environment env

val get_data : Env.t -> t

get_data () returns the data computed by the current processor