Module Basic.Debug

type flag

val d_warn : flag

val d_notice : flag

val register_flag : string -> flag

register_flag msg generates a new flag with error message msg

val new_flag : bool -> string -> flag

Same as register flag, but with a bool enbling the flag by default

val enable_flag : flag -> unit

Activates error messages associated to a flag

val disable_flag : flag -> unit

Desactivates error messages associated to a flag

val debug : flag -> ('a, Stdlib.Format.formatter, unit, unit) Stdlib.format4 -> 'a

debug f prints information on the standard error channel if the given flag f is currently active.

val debug_eval : flag -> (unit -> unit) -> unit

debug_eval f (fun () -> body evaluates body if the given flag f is currently active.