Module Kernel.Srcheck

val d_SR : Basic.Debug.flag

val srfuel : int Stdlib.ref

This parameters indicates how much time, substitution followed by reduction is applied on the type of the rhs of a rule to check that the rule preserves typing.

module SRChecker (R : Reduction.S) : sig ... end