Module Parsers.Preterm

PreTerms

This module defines structures representing terms before their scoping. That is to say before variables are scoped with De Bruijn indices

type preterm =

| PreType of Kernel.Basic.loc

| PreId of Kernel.Basic.loc * Kernel.Basic.ident

| PreQId of Kernel.Basic.loc * Kernel.Basic.name

| PreApp of preterm * preterm * preterm list

| PreLam of Kernel.Basic.loc * Kernel.Basic.ident * preterm option * preterm

| PrePi of Kernel.Basic.loc * Kernel.Basic.ident option * preterm * preterm

val pp_preterm : Stdlib.Format.formatter -> preterm -> unit

type prepattern =

| PCondition of preterm

| PPattern of Kernel.Basic.loc * Kernel.Basic.mident option * Kernel.Basic.ident * prepattern list

| PLambda of Kernel.Basic.loc * Kernel.Basic.ident * prepattern

| PJoker of Kernel.Basic.loc * prepattern list

val pp_prepattern : Stdlib.Format.formatter -> prepattern -> unit

type pdecl = (Kernel.Basic.loc * Kernel.Basic.ident) * preterm option

val pp_pdecl : Stdlib.Format.formatter -> pdecl -> unit

type pcontext = pdecl list

val pp_pcontext : Stdlib.Format.formatter -> pcontext -> unit

type prule = Kernel.Basic.loc * (Kernel.Basic.mident option * Kernel.Basic.ident) option * pdecl list * Kernel.Basic.mident option * Kernel.Basic.ident * prepattern list * preterm

val pp_prule : Stdlib.Format.formatter -> prule -> unit